pyproject.toml
[project]
name = "adk-from-idea-to-blueprint-agents   "
version = "0.1.0"
description = "Google ADK Agents PRD orchestration and more v1"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
"google-adk==0.5.0",
"pydantic==2.11.4",
"python-dotenv==1.1.0",
"google-genai==1.15.0",
"google-ai-generativelanguage==0.6.18",
"langchain==0.3.25",
"langchain-tavily==0.1.6",
"markdown==3.8",
"pdfkit==1.0.0",
"jinja2==3.1.6",
"filetype==1.2.0",
"langchain-google-genai==2.1.4",
]
